// vuejs repo/src/api/generated/models/UserDTOPaginationItemsApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserDTOPaginationItems } from './UserDTOPaginationItems'
export type UserDTOPaginationItemsApiResponse = {
  requestId?: string
  isSuccess?: boolean
  message?: string
  data?: UserDTOPaginationItems
}